move shared/lib/security to its main consumer
