+++ x-pack/platform/plugins/shared/osquery/cypress/cypress_base.config.ts
@@ -9,8 +9,8 @@ import { merge } from 'lodash';
 import path from 'path';
 import { load as loadYaml } from 'js-yaml';
 import { readFileSync } from 'fs';
-import type { YamlRoleDefinitions } from '@kbn/test-suites-serverless/shared/lib';
 import { samlAuthentication } from '@kbn/cypress-test-helper/src/auth/saml_auth';
+import type { YamlRoleDefinitions } from './lib';
 import { setupUserDataLoader } from './support/setup_data_loader_tasks';
 import { getFailedSpecVideos } from './support/filter_videos';
