+++ x-pack/platform/plugins/shared/osquery/cypress/lib/index.ts
