+++ x-pack/platform/plugins/shared/osquery/cypress/lib/types.ts
