+++ x-pack/platform/plugins/shared/osquery/cypress/support/e2e.ts
@@ -34,7 +34,7 @@ registerCypressGrep();
 import type { SecuritySolutionDescribeBlockFtrConfig } from '@kbn/cypress-test-helper/src/utils';
 import { login } from '@kbn/cypress-test-helper/src/auth/login';
 
-import type { LoadedRoleAndUser } from '@kbn/test-suites-serverless/shared/lib';
+import type { LoadedRoleAndUser } from '../lib';
 import type { ServerlessRoleName } from './roles';
 
 import { waitUntil } from '../tasks/wait_until';
